Removed generic placeholder for types
