+++ external-crates/move/crates/move-analyzer/src/analysis/parsing_analysis.rs
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
+    analysis::DefMap,
     symbols::{
         add_member_use_def,
         cursor::{CursorContext, CursorDefinition, CursorPosition},
         ignored_function,
         mod_defs::{AutoImportInsertionInfo, AutoImportInsertionKind, CallInfo, ModuleDefs},
         parsed_address, parsing_leading_and_mod_names_to_map_key, parsing_mod_def_to_map_key,
-        types::DefMap,
         use_def::{References, UseDef, UseDefMap},
     },
     utils::{loc_end_to_lsp_position_opt, loc_start_to_lsp_position_opt},
