+++ external-crates/move/crates/move-analyzer/src/analysis/typing_analysis.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
+    analysis::DefMap,
     compiler_info::CompilerInfo,
     symbols::{
         add_member_use_def,
@@ -9,7 +10,6 @@ use crate::{
         find_datatype,
         mod_defs::{MemberDefInfo, ModuleDefs},
         type_def_loc,
-        types::{DefMap, LocalDef},
         use_def::{References, UseDef, UseDefMap},
     },
     utils::{expansion_mod_ident_to_map_key, ignored_function, loc_start_to_lsp_position_opt},
@@ -31,7 +31,7 @@ use move_symbol_pool::Symbol;
 
 use im::OrdMap;
 use lsp_types::Position;
-use std::collections::BTreeMap;
+use std::{cmp, collections::BTreeMap};
 
 /// Data used during anlysis over typed AST
 pub struct TypingAnalysisContext<'a> {
@@ -64,6 +64,26 @@ pub struct TypingAnalysisContext<'a> {
     pub expression_scope: OrdMap<Symbol, LocalDef>,
     /// IDE Annotation Information from the Compiler
     pub compiler_info: &'a mut CompilerInfo,
+}
+
+/// Definition of a local (or parameter)
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct LocalDef {
+    /// Location of the definition
+    pub def_loc: Loc,
+    /// Type of definition
+    pub def_type: N::Type,
+}
+
+impl PartialOrd for LocalDef {
+    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+impl Ord for LocalDef {
+    fn cmp(&self, other: &Self) -> cmp::Ordering {
+        self.def_loc.cmp(&other.def_loc)
+    }
 }
 
 fn def_info_to_type_def_loc(
