+++ external-crates/move/crates/move-analyzer/src/code_action.rs
@@ -8,10 +8,10 @@ use crate::{
     },
     context::Context,
     symbols::{
+        Symbols,
         compilation::{CompiledPkgInfo, PrecomputedPkgInfo, get_compiled_pkg},
         cursor::{ChainInfo, CursorContext},
         runner::SymbolicatorRunner,
-        types::Symbols,
     },
     utils::loc_start_to_lsp_position_opt,
 };
