+++ external-crates/move/crates/move-analyzer/src/inlay_hints.rs
@@ -4,8 +4,8 @@
 use crate::{
     context::Context,
     symbols::{
-        def_info::DefInfo, ide_strings::type_to_ide_string, mod_defs::ModuleDefs,
-        requests::on_hover_markup, runner::SymbolicatorRunner, types::Symbols,
+        Symbols, def_info::DefInfo, ide_strings::type_to_ide_string, mod_defs::ModuleDefs,
+        requests::on_hover_markup, runner::SymbolicatorRunner,
     },
 };
 use lsp_server::{Message, Request, Response};
