+++ external-crates/move/crates/move-analyzer/src/symbols/cursor.rs
@@ -65,7 +65,7 @@ pub enum CursorDefinition {
 }
 
 //**************************************************************************************************
-// impls
+// Impls
 //**************************************************************************************************
 
 impl ChainInfo {
