+++ external-crates/move/crates/move-analyzer/src/symbols/mod_defs.rs
@@ -111,7 +111,7 @@ pub enum AutoImportInsertionKind {
 }
 
 //**************************************************************************************************
-// impls
+// Impls
 //**************************************************************************************************
 
 impl ModuleDefs {
