+++ external-crates/move/crates/move-analyzer/src/symbols/use_def.rs
@@ -48,7 +48,7 @@ pub struct UseLoc {
 pub type References = BTreeMap<Loc, BTreeSet<UseLoc>>;
 
 //**************************************************************************************************
-// impls
+// Impls
 //**************************************************************************************************
 
 impl Ord for UseDef {
